+++ CHANGELOG.md
@@ -18,6 +18,7 @@ v5.3
 **FieldConvert**
 - Added conditional to select the eNearestNeighbour method for 3D interpolation (!1335)
 - Fixed the output field names of WSS module of FieldConvert
+- Add VTU output using VTK library (high-order & multi-block options)
 
 **CompressibleFlowSolver**
 - Add additional parameters for the Isentropic Vortex equation system (!1323)